# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * GeoDataPlacefinder API — open-source geocoding powered by Overture Maps and DuckDB
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  /** Server status ("ok") */
  status: string;
  /** API version */
  version: string;
  /** Server uptime in seconds */
  uptime_s: number;
  /** Number of places loaded in the local DuckDB cache */
  cache_rows: number;
}

export interface Place {
  /** Overture Maps place identifier (e.g. "overture:place:abc123") */
  id?: string | null;
  /** Primary place name */
  name?: string | null;
  /** Freeform address string */
  address?: string | null;
  /** Latitude (WGS84) */
  lat: number;
  /** Longitude (WGS84) */
  lon: number;
  /** Data confidence score (0-1) */
  confidence?: number | null;
  /** Place category tags */
  categories?: string[] | null;
}

export type PlaceWithDistance = Place & {
  /** Distance from query coordinates in metres */
  distance_m: number;
};

export interface ErrorResponse {
  error: string;
}

export type SearchPlacesParams = {
/**
 * Free-form search query (e.g. "Cafe Paris New York")
 */
q?: string;
/**
 * Place name filter
 */
name?: string;
/**
 * City filter for structured queries
 */
city?: string;
/**
 * Postcode filter for structured queries
 */
postcode?: string;
/**
 * Place category/type filter (e.g. "cafe", "restaurant")
 */
type?: string;
/**
 * Maximum number of results (1-20)
 * @minimum 1
 * @maximum 20
 */
limit?: number;
};

export type ReversePlacesParams = {
/**
 * Latitude (WGS84)
 */
lat: number;
/**
 * Longitude (WGS84)
 */
lon: number;
/**
 * Maximum number of results (1-20)
 * @minimum 1
 * @maximum 20
 */
limit?: number;
};

export type NearestPlacesParams = {
/**
 * Latitude (WGS84)
 */
lat: number;
/**
 * Longitude (WGS84)
 */
lon: number;
/**
 * Optional place name filter
 */
name?: string;
/**
 * Maximum search radius in metres
 */
max_distance_m?: number;
};