# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: GeoDataPlacefinder API — open-source geocoding powered by Overture Maps and DuckDB
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /places/nearest:
    get:
      operationId: nearestPlaces
      tags: [geocoding]
      summary: Find nearest place to coordinates
      description: |
        Returns the single nearest place to given coordinates, optionally filtered by name.
        Useful for point-of-interest lookup and proximity search.
      parameters:
        - name: lat
          in: query
          description: Latitude (WGS84)
          required: true
          schema:
            type: number
            format: double
        - name: lon
          in: query
          description: Longitude (WGS84)
          required: true
          schema:
            type: number
            format: double
        - name: name
          in: query
          description: Optional place name filter
          required: false
          schema:
            type: string
        - name: max_distance_m
          in: query
          description: Maximum search radius in metres
          required: false
          schema:
            type: integer
            default: 5000
      responses:
        "200":
          description: Nearest place found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlaceWithDistance"
        "404":
          description: No place found within the specified radius
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "400":
          description: Missing or invalid lat/lon parameters
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
components:
  schemas:
...
    Place:
      type: object
      properties:
        id:
          type: string
          description: Overture Maps place identifier (e.g. "overture:place:abc123")
          nullable: true
        name:
          type: string
          description: Primary place name
          nullable: true
        address:
          type: string
          description: Freeform address string
          nullable: true
        lat:
          type: number
          format: double
          description: Latitude (WGS84)
        lon:
          type: number
          format: double
          description: Longitude (WGS84)
        confidence:
          type: number
          format: double
          description: Data confidence score (0-1)
          nullable: true
        categories:
          type: array
          items:
            type: string
          description: Place category tags
          nullable: true
      required:
        - lat
        - lon
    PlaceWithDistance:
      allOf:
        - $ref: "#/components/schemas/Place"
        - type: object
          properties:
            distance_m:
              type: number
              format: double
              description: Distance from query coordinates in metres
          required:
            - distance_m
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error