# lib/api-zod/src/generated/types/place.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * GeoDataPlacefinder API — open-source geocoding powered by Overture Maps and DuckDB
 * OpenAPI spec version: 0.1.0
 */

export interface Place {
  /** Overture Maps place identifier (e.g. "overture:place:abc123") */
  id?: string | null;
  /** Primary place name */
  name?: string | null;
  /** Freeform address string */
  address?: string | null;
  /** Latitude (WGS84) */
  lat: number;
  /** Longitude (WGS84) */
  lon: number;
  /** Data confidence score (0-1) */
  confidence?: number | null;
  /** Place category tags */
  categories?: string[] | null;
}